refactor(vela-ext): move test files from entrypoints/ to tests/
